optimized

+++ src/upgrade-self.c
@@ -1,74 +1,146 @@
 #include <math.h>
+#include <errno.h>
 #include <string.h>
+
+#include <limits.h>
 #include <sys/stat.h>
 
 #include "core/sysinfo.h"
 #include "core/self.h"
 #include "core/tar.h"
 #include "core/log.h"
-#include "core/cp.h"
-#include "autotools-setup.h"
+
+#include "main.h"
 
 int autotools_setup_upgrade_self(bool verbose) {
-    const char * const userHomeDir = getenv("HOME");
+    const char * const userHomeDIR = getenv("HOME");
 
-    if (userHomeDir == NULL) {
+    if (userHomeDIR == NULL) {
         return AUTOTOOLS_SETUP_ERROR_ENV_HOME_NOT_SET;
     }
 
-    size_t userHomeDirLength = strlen(userHomeDir);
-
-    if (userHomeDirLength == 0U) {
+    if (userHomeDIR[0] == '\0') {
         return AUTOTOOLS_SETUP_ERROR_ENV_HOME_NOT_SET;
     }
 
-    ////////////////////////////////////////////////////////////////
+    size_t homeDIRCapacity = strlen(userHomeDIR) + 18U;
+    char   homeDIR[homeDIRCapacity];
+
+    int ret = snprintf(homeDIR, homeDIRCapacity, "%s/.autotools-setup", userHomeDIR);
+
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
 
     struct stat st;
 
-    size_t   autotoolsSetupHomeDirLength = userHomeDirLength + 18U;
-    char     autotoolsSetupHomeDir[autotoolsSetupHomeDirLength];
-    snprintf(autotoolsSetupHomeDir, autotoolsSetupHomeDirLength, "%s/.autotools-setup", userHomeDir);
-
-    if (stat(autotoolsSetupHomeDir, &st) == 0) {
+    if (stat(homeDIR, &st) == 0) {
         if (!S_ISDIR(st.st_mode)) {
-            fprintf(stderr, "'%s\n' was expected to be a directory, but it was not.\n", autotoolsSetupHomeDir);
+            fprintf(stderr, "%s was expected to be a directory, but it was not.\n", homeDIR);
             return AUTOTOOLS_SETUP_ERROR;
         }
     } else {
-        if (mkdir(autotoolsSetupHomeDir, S_IRWXU) != 0) {
-            perror(autotoolsSetupHomeDir);
-            return AUTOTOOLS_SETUP_ERROR;
+        if (mkdir(homeDIR, S_IRWXU) != 0) {
+            if (errno != EEXIST) {
+                perror(homeDIR);
+                return AUTOTOOLS_SETUP_ERROR;
+            }
         }
     }
 
-    ////////////////////////////////////////////////////////////////
+    ///////////////////////////////////////////////////////////////////////////////////////
 
-    size_t   autotoolsSetupTmpDirLength = autotoolsSetupHomeDirLength + 5U;
-    char     autotoolsSetupTmpDir[autotoolsSetupTmpDirLength];
-    snprintf(autotoolsSetupTmpDir, autotoolsSetupTmpDirLength, "%s/tmp", autotoolsSetupHomeDir);
+    size_t runDIRCapacity = homeDIRCapacity + 5U;
+    char   runDIR[runDIRCapacity];
 
-    if (stat(autotoolsSetupTmpDir, &st) == 0) {
+    ret = snprintf(runDIR, runDIRCapacity, "%s/run", homeDIR);
+
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
+
+    if (lstat(runDIR, &st) == 0) {
         if (!S_ISDIR(st.st_mode)) {
-            fprintf(stderr, "'%s\n' was expected to be a directory, but it was not.\n", autotoolsSetupTmpDir);
-            return AUTOTOOLS_SETUP_ERROR;
+            if (unlink(runDIR) != 0) {
+                perror(runDIR);
+                return AUTOTOOLS_SETUP_ERROR;
+            }
+
+            if (mkdir(runDIR, S_IRWXU) != 0) {
+                if (errno != EEXIST) {
+                    perror(runDIR);
+                    return AUTOTOOLS_SETUP_ERROR;
+                }
+            }
         }
     } else {
-        if (mkdir(autotoolsSetupTmpDir, S_IRWXU) != 0) {
-            perror(autotoolsSetupTmpDir);
+        if (mkdir(runDIR, S_IRWXU) != 0) {
+            if (errno != EEXIST) {
+                perror(runDIR);
+                return AUTOTOOLS_SETUP_ERROR;
+            }
+        }
+    }
+
+    ////////////////////////////////////////////////////////////////////////////////////////////
+
+    size_t sessionDIRCapacity = runDIRCapacity + 20U;
+    char   sessionDIR[sessionDIRCapacity];
+
+    ret = snprintf(sessionDIR, sessionDIRCapacity, "%s/%d", runDIR, getpid());
+
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
+
+    if (lstat(sessionDIR, &st) == 0) {
+        if (S_ISDIR(st.st_mode)) {
+            ret = autotools_setup_rm_r(sessionDIR, false);
+
+            if (ret != AUTOTOOLS_SETUP_OK) {
+                return ret;
+            }
+
+            if (mkdir(sessionDIR, S_IRWXU) != 0) {
+                perror(sessionDIR);
+                return AUTOTOOLS_SETUP_ERROR;
+            }
+        } else {
+            if (unlink(sessionDIR) != 0) {
+                perror(sessionDIR);
+                return AUTOTOOLS_SETUP_ERROR;
+            }
+
+            if (mkdir(sessionDIR, S_IRWXU) != 0) {
+                perror(sessionDIR);
+                return AUTOTOOLS_SETUP_ERROR;
+            }
+        }
+    } else {
+        if (mkdir(sessionDIR, S_IRWXU) != 0) {
+            perror(sessionDIR);
             return AUTOTOOLS_SETUP_ERROR;
         }
     }
 
     ////////////////////////////////////////////////////////////////////////////////////////////
 
-    const char * const githubApiUrl = "https://api.github.com/repos/leleliu008/autotools-setup/releases/latest";
+    const char * githubApiUrl = "https://api.github.com/repos/leleliu008/autotools-setup/releases/latest";
 
-    size_t   githubApiResultJsonFilePathLength = autotoolsSetupTmpDirLength + 13U;
-    char     githubApiResultJsonFilePath[githubApiResultJsonFilePathLength];
-    snprintf(githubApiResultJsonFilePath, githubApiResultJsonFilePathLength, "%s/latest.json", autotoolsSetupTmpDir);
+    size_t githubApiResultJsonFilePathCapacity = sessionDIRCapacity + 13U;
+    char   githubApiResultJsonFilePath[githubApiResultJsonFilePathCapacity];
 
-    int ret = autotools_setup_http_fetch_to_file(githubApiUrl, githubApiResultJsonFilePath, verbose, verbose);
+    ret = snprintf(githubApiResultJsonFilePath, githubApiResultJsonFilePathCapacity, "%s/latest.json", sessionDIR);
+
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
+
+    ret = autotools_setup_http_fetch_to_file(githubApiUrl, githubApiResultJsonFilePath, verbose, verbose);
 
     if (ret != AUTOTOOLS_SETUP_OK) {
         return ret;
@@ -81,14 +153,18 @@ int autotools_setup_upgrade_self(bool verbose) {
         return AUTOTOOLS_SETUP_ERROR;
     }
 
-    char * latestVersion = NULL;
+    char * latestReleaseTagName = NULL;
+    size_t latestReleaseTagNameLength = 0U;
+
+    char   latestVersion[11] = {0};
+    size_t latestVersionLength = 0U;
 
     char * p = NULL;
 
-    char line[30];
+    char line[70];
 
     for (;;) {
-        p = fgets(line, 30, file);
+        p = fgets(line, 70, file);
 
         if (p == NULL) {
             if (ferror(file)) {
@@ -124,21 +200,27 @@ int autotools_setup_upgrade_self(bool verbose) {
                 }
             }
 
+            latestReleaseTagName = p;
+
             size_t n = 0;
-            char * q = p;
 
             for (;;) {
-                if (q[n] == '\0') {
+                if (p[n] == '\0') {
                     fprintf(stderr, "%s return invalid json.\n", githubApiUrl);
                     return AUTOTOOLS_SETUP_ERROR;
                 }
 
-                if (q[n] == '+') { // found right double quote
-                    q[n] = '\0';
-                    latestVersion = &q[0];
+                if (p[n] == '"') { // found right double quote
+                    p[n] = '\0';
+                    latestReleaseTagNameLength = n;
                     goto finalize;
                 } else {
                     n++;
+
+                    if (p[n] == '+') {
+                        latestVersionLength = n > 10 ? 10 : n;
+                        strncpy(latestVersion, p, latestVersionLength);
+                    }
                 }
             }
         }
@@ -147,16 +229,20 @@ int autotools_setup_upgrade_self(bool verbose) {
 finalize:
     fclose(file);
 
-    printf("latestVersion=%s\n", latestVersion);
+    printf("latestReleaseTagName=%s\n", latestReleaseTagName);
 
-    if (latestVersion == NULL) {
+    if (latestReleaseTagName == NULL) {
         fprintf(stderr, "%s return json has no tag_name key.\n", githubApiUrl);
         return AUTOTOOLS_SETUP_ERROR;
     }
 
-    size_t latestVersionCopyLength = strlen(latestVersion) + 1U;
-    char   latestVersionCopy[latestVersionCopyLength];
-    strncpy(latestVersionCopy, latestVersion, latestVersionCopyLength);
+    if (latestVersion[0] == '\0') {
+        fprintf(stderr, "%s return invalid json.\n", githubApiUrl);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
+
+    char    latestVersionCopy[latestVersionLength + 1U];
+    strncpy(latestVersionCopy, latestVersion, latestVersionLength + 1U);
 
     char * latestVersionMajorStr = strtok(latestVersionCopy, ".");
     char * latestVersionMinorStr = strtok(NULL, ".");
@@ -214,19 +300,35 @@ finalize:
         return AUTOTOOLS_SETUP_ERROR;
     }
 
-    size_t latestVersionLength = strlen(latestVersion);
+    size_t tarballFileNameCapacity = latestVersionLength + strlen(osType) + strlen(osArch) + 26U;
+    char   tarballFileName[tarballFileNameCapacity];
 
-    size_t   tarballFileNameLength = latestVersionLength + strlen(osType) + strlen(osArch) + 26U;
-    char     tarballFileName[tarballFileNameLength];
-    snprintf(tarballFileName, tarballFileNameLength, "autotools-setup-%s-%s-%s.tar.xz", latestVersion, osType, osArch);
+    ret = snprintf(tarballFileName, tarballFileNameCapacity, "autotools-setup-%s-%s-%s.tar.xz", latestVersion, osType, osArch);
 
-    size_t   tarballUrlLength = tarballFileNameLength + latestVersionLength + 66U;
-    char     tarballUrl[tarballUrlLength];
-    snprintf(tarballUrl, tarballUrlLength, "https://github.com/leleliu008/autotools-setup/releases/download/%s/%s", latestVersion, tarballFileName);
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
 
-    size_t   tarballFilePathLength = autotoolsSetupTmpDirLength + tarballFileNameLength + 2U;
-    char     tarballFilePath[tarballFilePathLength];
-    snprintf(tarballFilePath, tarballFilePathLength, "%s/%s", autotoolsSetupTmpDir, tarballFileName);
+    size_t tarballUrlCapacity = tarballFileNameCapacity + strlen(latestReleaseTagName) + 66U;
+    char   tarballUrl[tarballUrlCapacity];
+
+    ret = snprintf(tarballUrl, tarballUrlCapacity, "https://github.com/leleliu008/autotools-setup/releases/download/%s/%s", latestReleaseTagName, tarballFileName);
+
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
+
+    size_t tarballFilePathCapacity = sessionDIRCapacity + tarballFileNameCapacity + 2U;
+    char   tarballFilePath[tarballFilePathCapacity];
+
+    ret = snprintf(tarballFilePath, tarballFilePathCapacity, "%s/%s", sessionDIR, tarballFileName);
+
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
 
     ret = autotools_setup_http_fetch_to_file(tarballUrl, tarballFilePath, verbose, verbose);
 
@@ -236,19 +338,21 @@ finalize:
 
     //////////////////////////////////////////////////////////////////////////////////
 
-    size_t   tarballExtractDirLength = tarballFilePathLength + 3U;
-    char     tarballExtractDir[tarballExtractDirLength];
-    snprintf(tarballExtractDir, tarballExtractDirLength, "%s.d", tarballFilePath);
-
-    ret = tar_extract(tarballExtractDir, tarballFilePath, 0, verbose, 1);
+    ret = tar_extract(sessionDIR, tarballFilePath, 0, verbose, 1);
 
     if (ret != 0) {
         return abs(ret) + AUTOTOOLS_SETUP_ERROR_ARCHIVE_BASE;
     }
 
-    size_t   upgradableExecutableFilePathLength = tarballExtractDirLength + 21U;
-    char     upgradableExecutableFilePath[upgradableExecutableFilePathLength];
-    snprintf(upgradableExecutableFilePath, upgradableExecutableFilePathLength, "%s/bin/autotools-setup", tarballExtractDir);
+    size_t upgradableExecutableFilePathCapacity = sessionDIRCapacity + 20U;
+    char   upgradableExecutableFilePath[upgradableExecutableFilePathCapacity];
+
+    ret = snprintf(upgradableExecutableFilePath, upgradableExecutableFilePathCapacity, "%s/bin/autotools-setup", sessionDIR);
+
+    if (ret < 0) {
+        perror(NULL);
+        return AUTOTOOLS_SETUP_ERROR;
+    }
 
     char * selfRealPath = self_realpath();
 
@@ -258,21 +362,29 @@ finalize:
         goto finally;
     }
 
-    if (unlink(selfRealPath) != 0) {
-        perror(selfRealPath);
-        ret = AUTOTOOLS_SETUP_ERROR;
-        goto finally;
-    }
+    if (rename(upgradableExecutableFilePath, selfRealPath) != 0) {
+        if (errno == EXDEV) {
+            if (unlink(selfRealPath) != 0) {
+                perror(selfRealPath);
+                ret = AUTOTOOLS_SETUP_ERROR;
+                goto finally;
+            }
 
-    if (copy_file(upgradableExecutableFilePath, selfRealPath) < 0) {
-        perror(NULL);
-        ret = AUTOTOOLS_SETUP_ERROR;
-        goto finally;
-    }
+            ret = autotools_setup_copy_file(upgradableExecutableFilePath, selfRealPath);
 
-    if (chmod(selfRealPath, S_IRWXU) != 0) {
-        perror(selfRealPath);
-        ret = AUTOTOOLS_SETUP_ERROR;
+            if (ret != AUTOTOOLS_SETUP_OK) {
+                goto finally;
+            }
+
+            if (chmod(selfRealPath, S_IRWXU) != 0) {
+                perror(selfRealPath);
+                ret = AUTOTOOLS_SETUP_ERROR;
+            }
+        } else {
+            perror(selfRealPath);
+            ret = AUTOTOOLS_SETUP_ERROR;
+            goto finally;
+        }
     }
 
 finally:
